requirements.txt
python-docx==0.8.11
flask==2.3.3
requests==2.31.0
gunicorn==21.2.0